Update 633. Sum of Square Numbers.java

+++ Medium_LeetCode/633. Sum of Square Numbers.java
@@ -1,8 +1,8 @@
 class Solution {
     public boolean judgeSquareSum(int c) {
-        int left = 0, right = (int)Math.sqrt(c);
+        long left = 0, right = (long) Math.sqrt(c);
         while (left <= right) {
-            int cur = left * left + right * right;
+            long cur = left * left + right * right;
             if (cur < c) {
                 left++;
             } else if (cur > c) {
